refactor(anthropic): extract pure convertMessages/convertTools to anthropic-convert.js
PR #155 review: anthropic-counter.countMessagesAsync was importing convertMessages from anthropic-client.js, which loads '../logger' at the top level (and logger requires 'vscode'). In a plain-Node environment, the async exact-count path would crash on import and silently fall back to the heuristic, defeating the modular counter design.

Moved the two pure helpers into src/api/anthropic-convert.js (no side effects, no vscode/logger/SDK deps). anthropic-client.js re-exports them via require('./anthropic-convert'), and anthropic-counter.js imports directly from the pure module. Verified by stubbing require('vscode') to throw and successfully calling convertMessages.
